Use already existing dtype_string function
