+++ src/common/utils.h
@@ -365,30 +365,6 @@ inline bool ContainsStorageType(const std::vector<int>& ndstypes,
   return false;
 }
 
-inline std::string dtype_string(const int dtype) {
-  switch (dtype) {
-    case mshadow::kFloat32:
-      return "float";
-    case mshadow::kFloat64:
-      return "double";
-    case mshadow::kFloat16:
-      return "half";
-    case mshadow::kUint8:
-      return "unsigned char";
-    case mshadow::kInt8:
-      return "char";
-    case mshadow::kInt32:
-      return "int";
-    case mshadow::kInt64:
-      return "long long";
-    case mshadow::kBool:
-      return "bool";
-    default:
-      LOG(FATAL) << "Unknown type enum " << dtype;
-  }
-  return "unknown";
-}
-
 /*! \brief get string representation of dispatch_mode */
 inline std::string dispatch_mode_string(const DispatchMode x) {
   switch (x) {
